refactor for ecdsa
