+++ native/src/protocols/two_party_ecdsa/party1.rs
@@ -3,7 +3,7 @@ extern crate server_lib;
 use neon::prelude::*;
 use self::server_lib::server::*;
 
-pub fn p1_launch_server(mut cx: FunctionContext) -> JsResult<JsString> {
+pub fn launch_server(mut cx: FunctionContext) -> JsResult<JsString> {
     get_server().launch();
     Ok(cx.string("success"))
 }
